+++ financepy/__init__.py
@@ -2,7 +2,7 @@ __version__ = "1.1.0"
 
 CR = "\n"
 S =  "#############################################################" + CR
-S += "#  FINANCEPY Version 1.1.0 - Built on 20 Aug 2026 at 22:16  #" + CR
+S += "#  FINANCEPY Version 1.1.0 - Built on 21 Aug 2026 at 15:41  #" + CR
 S += "#  This software is distributed FREE AND WITHOUT WARRANTY   #" + CR
 S += "#  Report issues at https://github.com/domokane/FinancePy   #" + CR
 S += "#############################################################"